- refactor: PotEvent 인터페이스 구조 변경, dto -> schema 이름 변경
