+++ src/pot/event/v1/dto/pot-event.chat.v1.dto.ts
