+++ src/pot/event/v1/dto/pot-event.user-in.v1.dto.ts
