+++ src/pot/event/v1/dto/pot-event.user-kick.v1.dto.ts
